snake: add a loading screen
